requirements.txt
blinker==1.7.0
click==8.1.7
cs50==9.3.4
Deprecated==1.2.14
Flask==3.0.3
Flask-Limiter==3.5.1
greenlet==3.0.3
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
limits==3.10.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.0
Pygments==2.17.2
python-dotenv==1.0.1
rich==13.7.1
SQLAlchemy==2.0.29
sqlparse==0.5.0
termcolor==2.4.0
typing_extensions==4.11.0
Werkzeug==3.0.2
wrapt==1.16.0
blinker==1.7.0
click==8.1.7
cs50==9.3.4
Deprecated==1.2.14
Flask==3.0.3
Flask-Limiter==3.5.1
greenlet==3.0.3
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
limits==3.10.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.0
Pygments==2.17.2
python-dotenv==1.0.1
rich==13.7.1
SQLAlchemy==2.0.29
sqlparse==0.5.0
termcolor==2.4.0
typing_extensions==4.11.0
Werkzeug==3.0.2
wrapt==1.16.0
